clean up editor code structure
