+++ src/editor/index.tsx
@@ -1,0 +1,12 @@
+export * from "./base";
+export * from "./functions";
+export * from "./graphql";
+export * from "./icons";
+export * from "./keybinds";
+export * from "./keymaps";
+export * from "./records";
+export * from "./selection";
+export * from "./surrealql";
+export * from "./tables";
+export * from "./theme";
+export * from "./variables";
